Factoring out forward decls into TaskTypes.h to minimize header dependiencies.
