+++ include/Task.h
@@ -18,6 +18,9 @@
 //--- User configuration header ---//
 #include "TasksConfig.h"
 
+//--- C++17/C++20 Compatibility ---//
+#include "Private/TasksCommonPrivate.h"
+
 //--- Debug Macros ---//
 #if SQUID_ENABLE_TASK_DEBUG
 /// @ingroup Tasks
@@ -45,40 +48,13 @@
 
 NAMESPACE_SQUID_BEGIN
 
+#include "TaskTypes.h"
+
 /// @addtogroup Tasks
 /// @{
 
-//--- Task Reference Type ---//
-enum class eTaskRef /// Whether a handle references a task using a strong or weak reference
-{
-	Strong, ///< Handle will keep the task alive (so long as there exists a valid Resumable handle)
-	Weak, ///< Handle will not the task alive
-};
-
-//--- Task Resumable Type ---//
-enum class eTaskResumable /// Whether a handle can be resumed (all live tasks have exactly one resumable handle and 0+ non-resumable handles)
-{
-	Yes, ///< Handle is resumable
-	No, ///< Handle is not resumable
-};
-
-//--- Task Status ---//
-enum class eTaskStatus /// Status of a task (whether it is currently suspended or done)
-{
-	Suspended, ///< Task is currently suspended
-	Done, ///< Task has terminated and coroutine frame has been destroyed
-};
-
 //--- tTaskCancelFn ---//
 using tTaskCancelFn = std::function<bool()>; ///< CancelIf/StopIf condition function type
-
-// Forward declarations
-template <typename tRet, eTaskRef RefType, eTaskResumable Resumable>
-class Task; /// Templated handle type (defaults to <void, Strong, Resumable>)
-template <typename tRet = void>
-using TaskHandle = Task<tRet, eTaskRef::Strong, eTaskResumable::No>; ///< Non-resumable handle that holds a strong reference to a task
-using WeakTask = Task<void, eTaskRef::Weak, eTaskResumable::Yes>; ///< Resumable handle that holds a weak reference to a task (always void return type)
-using WeakTaskHandle = Task<void, eTaskRef::Weak, eTaskResumable::No>; ///< Non-resumable handle that holds a weak reference to a task (always void return type)
 
 /// @} end of addtogroup Tasks
 
@@ -199,7 +175,7 @@ struct GetStopContext
 /// @tparam tRet Return type of the underlying coroutine (can be void if the coroutine does not co_return a value)
 /// @tparam RefType Whether this handle holds a strong or weak reference to the underlying coroutine
 /// @tparam Resumable Whether this handle can be used to resume the underlying coroutine
-template <typename tRet = void, eTaskRef RefType = eTaskRef::Strong, eTaskResumable Resumable = eTaskResumable::Yes>
+template <typename tRet, eTaskRef RefType, eTaskResumable Resumable>
 class Task
 {
 public:
@@ -594,7 +570,7 @@ private:
 /// Here is an example Squid::GetTime() function implementation from within the ```main.cpp``` file of a sample project:
 ///	
 ///	~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~.cpp
-///	NAMESPACE_SQUID_BEGIN
+///	
 ///	tTaskTime GetTime()
 ///	{
 ///		return (tTaskTime)TimeSystem::GetTime();
