+++ include/TokenList.h
@@ -66,6 +66,9 @@
 //--- User configuration header ---//
 #include "TasksConfig.h"
 
+//--- C++17/C++20 Compatibility ---//
+#include "Private/TasksCommonPrivate.h"
+
 NAMESPACE_SQUID_BEGIN
 
 template <typename T = void>
